Move ReadResourceSystem into read_resource module
